+++ src/kernel/syscall_internal.h
@@ -163,6 +163,7 @@ void h_ramfs_create(struct interrupt_frame64 *r);
 void h_block_read(struct interrupt_frame64 *r);
 void h_block_write(struct interrupt_frame64 *r);
 void h_storage_info(struct interrupt_frame64 *r);
+void h_storage_device(struct interrupt_frame64 *r);
 void h_storage_format(struct interrupt_frame64 *r);
 void h_register_fs_server(struct interrupt_frame64 *r);
 void h_connect_fs_server(struct interrupt_frame64 *r);